GerritStats: initial commit
This command line tool processes Gerrit code review output and creates
statistics out of it. This can be interesting just for fun, but also for
analysing how effective code reviews are among the team, what kind of
outliers can be found. The original reason for implementing this tool was
to study team dynamics and how well code reviews are implemented in
practice in teams.

The tools supports both CSV and plain text output. It can use both cached
gerrit query statistics files, as well as fetch data from the server and
use that data directly.