Replace custom command line parsing with JCommander
Done to reduce the parsing code and make it straightforward to make
modifications. There's a bit too many arguments now, and one way to
solve it would be to split the tool into two: a data downloader and
a stats parser. To unify the code of two tools, a command line parser
is used as a pre-step.