Add basic support for processing multiple files at once
Many projects have multiple repositories where same engineers contribute
to, so it's useful to generate stats for such situations as well.

There's a lot of room to improve this - for large multirepository projects,
the file lists are now way too long, and there's also issues with users
that have committed from multiple email addresses if no filtering is
applied. This will require further configuration switches in the future.