Index: limit filename output to 30 files and suppress output after that
If you pass e.g. 200 files to GerritStats, index.html's header would list
all the filenames. This is not very meaningful and clutters the output.
Now, for over 20 filenames, only the first and last filenames are listed,
and the total number of files is shown. It's possible to improve this by
e.g. making the text element expand all the filenames if necessary.