Html index page: add sortable table of most of the data
Uses tablesorter to easily sort the columns. Currently, the time in review
column is incorrectly sorted; this will be fixed in a separate commit.

Introduces numeral.js and jquery as dependencies for HTML as well. Numeral
is added to allow frontend to make display decisions as it should, and
jquery is a dependency of tablesorter.