GerritDownloader: pass in output dir instead of file
If no projects are specified, first get the project listing, then
download each project in its own .json file. This allows for easier use
of GerritStats tool later, as it support multiple projects, and reduces
need for writing custom scripts.

A set of wrapper shell scripts are included in the root to start up all
the tools, once they're built.