Read all information on person profiles from .js file
This is a non-behavior changing commit and enables more dynamic processing
of the data, with minimal need for command-line usage.

Currently, jquery is used for filling in the template, but as the data
generation (Java side) becomes fully separated from the presentation,
the presentation layer could be changed to use some other client-side
technology.

A dependency is added to moment.js to parse dates on client-side.