Make proximity graph more dynamic by allowing selection
The checkboxes are now functional, and allow for selection per user.
This allows to better indicate which users interact with each other,
without having to recompile the data with the tool.

An additional filter/mapping step is needed for links, so that:
they're passed to d3 using references to the node objects, using the
same node array that's passed to d3 as well. The indices are also not
used when generating data from velocity templates, as the references
are easier to work with when debugging.