Replace remaining velocity usages from index.vm
Also, remove the Velocity dependency entirely. Now the frontend is
purely JS-based, which is perhaps a bit slower as more data
processing is done on frontend, but which also allows for faster
development (as only HTML/js needs to be changed in many cases),
and for improved user experience as data can be analyzed without
having to regenerate the data with different options.