Show self-reviewed commit count, and count of commits now in review
In this context, self-reviews are considered to be self- +2'd commits,
with no other +2 approvals at any point in the changeset's history. This
approach avoids false positives where the user e.g. had to rebase and
lost the review although the rebase was trivial.

Both the index and the profile page show this number.