Refer to HTML local storage through a data set key
If multiple sets of data are fetched one after the other from Gerrit
server, the changes in data are not reflected in local storage as long
as the paths stay the same. To fix this, create a hashcode of all the
filenames, and use that as a reference. The solution could be still
improved by hashing the files, and then using a single hash representing
that file group to key into the storage.