GerritDownloader Make the generated .json files actual JSON
Before, the files were json only line-by-line; each line was a valid
json object, but the file as a whole as not. This prevented use of
common JSON tools, and making interoperability in general more difficult.

GerritStats implements not a parser for both old and new file formats, so
that already downloaded data files don't have to be redownloaded.
File format versioning is included to prevent any future compatibility
issues.