GerritStats: provide option to anonymize data
This can be useful for demo purposes, or in the case someone feels
offended by statistic being collected.

This feature removes all the visible references to an organization
and its members. To achieve this end, it also lorem ipsumizes all
review comments, and generates random, "techy" commit messages
and file paths to hide what is actually being done. The statistics
themselves are untouched.

This commit makes Java 8 a requirement for this project.