Add warning message to UI if gerrit version is not at least 2.9
If the server uses e.g. Gerrit 2.7, which e.g. Qt still uses, the data
on who was added as a reviewer is not there. Since this tool is heavily
based on this data, the output will be largely incorrect. Let the user
know using an alert box that their Gerrit version is too old; there's no
other fix for this.