Always use a unique project name when anonymizing
It's possible that for large organizations with a lot of projects,
the project names could collide, and lead to incorrect presentation of
data for that project. Now, there's tracking for used names, and in the
case really large amount of projects are processed (exceeding about 600),
a suffix is added to the generated name to guarantee uniqueness.