Rewrite the UI with modern Javascript (React / ES2015)
This rewrite is done in order to modernize the frontend of the tool's
HTML output. It uses all the cool things like webpack, React, Babel,
immutable and many other JS libs to nicely encapsulate and give structure
to the UI.

This also eases some parts of the UI development: simply running
`npm run webpack-watch` after building the data files automatically
recompiles the .jsx to .js, leaning itself to a pretty fast development
experience.

This does not require a node server to display the HTML output.

This huge commit also fixes some bugs, like not showing the right
amount of approvals per user, and improves the UI for the overview page
by moving the team graph to its own widget at the bottom of the page,
instead of overlaying it on top of the table.