Improve the command line tool to take two build systems into consideration
Since there's now both gradle and npm scripts/webpack as build systems,
there needs to be a tool to run the output into HTML without having to
run several commands. Now, gerrit_stats.sh does this. This is a bit
confusing if the user would like to have textual input only, so to that
end, the text-only command line tool will be likely removed from the
codebase later.