Remove non-HTML output from GerritStats
The HTML tool does everything that the command-line tool does, so this
should really not be necessary.

If someone really needs this, this could be reverted and separated into
its own tool, so that when running ./gerrit_stats.sh, the help would not
contain any reference to CSV or plaintext options, or other than
per-person data processing types.