Streamline build system for HTML output
Previously, the ids were output as an immutable map that had to be
imported using ES2015 syntax. To make that work (as it required
Babel), the data was generated into src/main/frontend directory.
This however is not where the data should live, as all the rest of
it is simply loaded, and it would unnecessarily lengthen the
webpack build time. Now, the development flow is like this:

1) ./gradlew assemble (or use e.g. IntelliJ to build)
2) cd GerritStats && npm run generateData -- -f your_data_file.json
3) npm run webpack-watch
4) Open the output in GerritStats/out-html and start development.

I noticed that there's some issues with moowork.node plugin that
prevent it from working in some cases. git clean -fxd fixes those
issues.